Fase 2: proteger el rol de pagos ajeno y usar Decimal para el dinero en RRHH
Hallazgo #4: SalaryPrintReceiptView imprimía cualquier rol de pagos por
su pk en la URL, sin revisar si pertenecía al usuario que lo pedía -
cualquier empleado logueado podía ver el sueldo de otro cambiando el
número. Ahora se verifica que, si quien pide el PDF es un empleado, el
rol de pagos sea el suyo propio (un administrador sigue pudiendo
imprimir cualquiera, como antes). Verificado: el empleado 1 puede
imprimir el suyo, no puede imprimir el del empleado 2 (redirige en vez
de devolver el PDF).

De paso se encontró y corrigió un bug real: is_employee() se usaba en
esa misma vista pero nunca se había definido en el modelo User -hubiera
lanzado un error si alguna vez se ejecutaba ese camino-. Se agregó
siguiendo el mismo patrón que is_client() ya existente.

También se cerró la exportación de empleados a Excel
(EmployeeExportExcelView), que solo pedía sesión iniciada y no
verificaba ningún permiso, pese a incluir la remuneración de todos los
empleados en el archivo. Ahora exige el mismo permiso view_employee que
ya protege el listado (ambos exponían los mismos datos, solo que uno sí
estaba protegido y el otro no). Verificado: un empleado sin ese permiso
ya no puede descargarlo, un administrador sigue pudiendo.

Por último, Employee.remuneration, SalaryDetail.income/expenses/
total_amount y SalaryHeadings.valor pasan de FloatField a DecimalField
(mismo criterio que ya se usa en Facturación), evitando el mismo tipo de
error de redondeo que se corrigió antes en el cálculo del IVA. No había
datos existentes en estas tablas en este ambiente, así que la migración
no tuvo que convertir nada.

+++ core/rrhh/models.py
@@ -44,7 +44,7 @@ class Employee(models.Model):
     hiring_date = models.DateField(default=datetime.now, verbose_name='Fecha de ingreso')
     position = models.ForeignKey(Position, on_delete=models.PROTECT, verbose_name='Cargo')
     area = models.ForeignKey(Area, on_delete=models.PROTECT, verbose_name='Area')
-    remuneration = models.FloatField(default=0.00, verbose_name='Remuneración')
+    remuneration = models.DecimalField(max_digits=9, decimal_places=2, default=0.00, verbose_name='Remuneración')
 
     def __str__(self):
         return self.get_full_name()
@@ -156,9 +156,9 @@ class Salary(models.Model):
 class SalaryDetail(models.Model):
     salary = models.ForeignKey(Salary, on_delete=models.CASCADE)
     employee = models.ForeignKey(Employee, on_delete=models.PROTECT, verbose_name='Empleado')
-    income = models.FloatField(default=0.00)
-    expenses = models.FloatField(default=0.00)
-    total_amount = models.FloatField(default=0.00)
+    income = models.DecimalField(max_digits=9, decimal_places=2, default=0.00)
+    expenses = models.DecimalField(max_digits=9, decimal_places=2, default=0.00)
+    total_amount = models.DecimalField(max_digits=9, decimal_places=2, default=0.00)
 
     def __str__(self):
         return self.employee.user.names
@@ -197,7 +197,7 @@ class SalaryHeadings(models.Model):
     salary_detail = models.ForeignKey(SalaryDetail, on_delete=models.CASCADE)
     headings = models.ForeignKey(Headings, on_delete=models.PROTECT)
     cant = models.IntegerField(default=0)
-    valor = models.FloatField(default=0.00)
+    valor = models.DecimalField(max_digits=9, decimal_places=2, default=0.00)
 
     def __str__(self):
         return self.salary_detail.employee.user.names
